State: Dynamically load AT status refetch
